test(frontend): remove 3 duplicate JSX test files (#35590)
